+++ src/commands/createFunction/FunctionListStep.ts
@@ -23,7 +23,8 @@ import { DotnetNamespaceStep } from './dotnetSteps/DotnetNamespaceStep';
 import { IFunctionWizardContext } from './IFunctionWizardContext';
 import { JavaFunctionCreateStep } from './javaSteps/JavaFunctionCreateStep';
 import { JavaFunctionNameStep } from './javaSteps/JavaFunctionNameStep';
-import { OpenAPICreateStep, OpenAPIGetSpecificationFileStep } from './openAPISteps/OpenAPICreateStep';
+import { OpenAPICreateStep } from './openAPISteps/OpenAPICreateStep';
+import { OpenAPIGetSpecificationFileStep } from './openAPISteps/OpenAPIGetSpecificationFileStep';
 import { ScriptFunctionCreateStep } from './scriptSteps/ScriptFunctionCreateStep';
 import { ScriptFunctionNameStep } from './scriptSteps/ScriptFunctionNameStep';
 import { TypeScriptFunctionCreateStep } from './scriptSteps/TypeScriptFunctionCreateStep';
